camera and other refractoring

+++ scripts/ar_object_importer.py
@@ -1,7 +1,7 @@
 """Module that contains the ArObjectImporter"""
 
 from ar_track_alvar_msgs.msg import AlvarMarkers
-from ar_track_alvar_msgs.msg import AlvarMarker
+from geometry_msgs.msg import Pose
 import rospy
 
 
@@ -20,6 +20,7 @@ class ArObjectImporter():
     self.is_visible = False
     self.max_ticks_since_seen = 10
     self.ticks_since_seen = self.max_ticks_since_seen + 1
+    self.last_pose = Pose()
 
     rospy.Subscriber("/ar_pose_marker", AlvarMarkers, self.__update)
 
@@ -33,7 +34,8 @@ class ArObjectImporter():
       if (len(alvar_markers.markers) > 0):
         self.ticks_since_seen = 0
         self.is_visible = True
-        self.moveit_object.set_pose(alvar_markers.markers[0].pose.pose)
+        self.last_pose = alvar_markers.markers[0].pose.pose
+        self.moveit_object.set_pose(self.last_pose)
       else:
         self.ticks_since_seen += 1
         if (self.ticks_since_seen > self.max_ticks_since_seen and
@@ -44,5 +46,6 @@ class ArObjectImporter():
   def stop(self):
     """Stops updating and removes the object"""
     self.is_stopped = True
-    self.is_visible = False
-    self.moveit_object.delete()
+    if (not self.is_visible):
+      self.moveit_object.delete()
+      self.is_visible = False
